migration of utility functions used by Math.h to headeronly
