+++ test/cpp/aoti_abi_check/test_math.cpp
@@ -1,6 +1,7 @@
 #include <gtest/gtest.h>
 
 #include <c10/util/generic_math.h>
+#include <torch/headeronly/util/BFloat16-math.h>
 #include <torch/headeronly/util/NumericUtils.h>
 #include <cmath>
 #include <limits>
@@ -60,6 +61,25 @@ TEST(TestMath, TestIsInf) {
   EXPECT_FALSE(_isinf(torch::headeronly::Float8_e4m3fn(1.0f)));
   EXPECT_FALSE(_isinf(torch::headeronly::Float8_e5m2fnuz(1.0f)));
   EXPECT_FALSE(_isinf(torch::headeronly::Float8_e4m3fnuz(1.0f)));
+}
+
+TEST(TestMath, TestIsReducedFloatingPoint) {
+  using torch::headeronly::BFloat16;
+  using torch::headeronly::Half;
+  using torch::headeronly::is_reduced_floating_point;
+  using torch::headeronly::is_reduced_floating_point_v;
+
+  static_assert(is_reduced_floating_point_v<Half>);
+  static_assert(is_reduced_floating_point_v<BFloat16>);
+  static_assert(!is_reduced_floating_point_v<float>);
+  EXPECT_TRUE(is_reduced_floating_point<Half>::value);
+  EXPECT_TRUE(is_reduced_floating_point<BFloat16>::value);
+  EXPECT_FALSE(is_reduced_floating_point<double>::value);
+
+  // Exercise the std:: overloads that dispatch on
+  // is_reduced_floating_point_v.
+  EXPECT_NEAR(static_cast<float>(std::acos(Half(0.5f))), 1.0472f, 1e-3);
+  EXPECT_NEAR(static_cast<float>(std::sqrt(BFloat16(4.0f))), 2.0f, 1e-2);
 }
 
 } // namespace aot_inductor
